move all llm dependency methods to agent
