+++ swissknife/modules/gui/qt_ui.py
@@ -114,7 +114,7 @@ class ChatWindow(QMainWindow, Observer):
 
         # Create the status indicator (showing current agent and model)
         self.status_indicator = QLabel(
-            f"Agent: {self.message_handler.agent.name} | Model: {self.message_handler.agent.llm.model}"
+            f"Agent: {self.message_handler.agent.name} | Model: {self.message_handler.agent.get_model()}"
         )
         self.status_indicator.setStyleSheet(
             """
@@ -651,7 +651,7 @@ class ChatWindow(QMainWindow, Observer):
         # self.display_response_chunk(response)
 
         # Calculate cost
-        total_cost = self.message_handler.agent.llm.calculate_cost(
+        total_cost = self.message_handler.agent.calculate_usage_cost(
             input_tokens, output_tokens
         )
 
@@ -1279,7 +1279,7 @@ class ChatWindow(QMainWindow, Observer):
         try:
             # Format the messages for display
             debug_info = json.dumps(self.message_handler.agent.history, indent=2)
-        except Exception as e:
+        except Exception as _:
             debug_info = str(self.message_handler.agent.history)
         # Add as a system message
         self.add_system_message(f"DEBUG INFO:\n\n```json\n{debug_info}\n```")
@@ -1287,7 +1287,7 @@ class ChatWindow(QMainWindow, Observer):
         try:
             # Format the messages for display
             debug_info = json.dumps(self.message_handler.streamline_messages, indent=2)
-        except Exception as e:
+        except Exception as _:
             debug_info = str(self.message_handler.streamline_messages)
         # Add as a system message
         self.add_system_message(f"DEBUG INFO:\n\n```json\n{debug_info}\n```")
@@ -1351,7 +1351,7 @@ class ChatWindow(QMainWindow, Observer):
     @Slot(str, object)
     def handle_event(self, event: str, data: Any):
         if event == "response_chunk":
-            chunk, assistant_response = data
+            _, assistant_response = data
             if assistant_response.strip():
                 self.display_response_chunk(assistant_response)
         elif event == "error":
@@ -1421,7 +1421,7 @@ class ChatWindow(QMainWindow, Observer):
         elif event == "agent_changed":
             self.add_system_message(f"Switched to {data} agent")
             self.status_indicator.setText(
-                f"Agent: {data} | Model: {self.message_handler.agent.llm.model}"
+                f"Agent: {data} | Model: {self.message_handler.agent.get_model()}"
             )
         elif event == "model_changed":
             self.add_system_message(f"Switched to {data['name']} ({data['id']})")
@@ -1431,7 +1431,7 @@ class ChatWindow(QMainWindow, Observer):
         elif event == "agent_changed_by_transfer":
             self.add_system_message(f"Transfered to {data} agent")
             self.status_indicator.setText(
-                f"Agent: {data} | Model: {self.message_handler.agent.llm.model}"
+                f"Agent: {data} | Model: {self.message_handler.agent.get_model()}"
             )
             # Reset the current response bubble so the next agent message starts in a new bubble
             self.current_response_bubble = None
